+++ paper/src/main/kotlin/dev/nyon/magnetic/Listeners.kt
@@ -185,7 +185,8 @@ object Listeners {
 
     private val cooldowns: Map<Config.FullInventoryAlert.Alert, MutableMap<UUID, Instant>> = mapOf(
         config.fullInventoryAlert.soundAlert to mutableMapOf(),
-        config.fullInventoryAlert.textAlert to mutableMapOf()
+        config.fullInventoryAlert.textAlert to mutableMapOf(),
+        config.fullInventoryAlert.titleAlert to mutableMapOf()
     )
     private fun tickInventoryAlert(player: Player) {
         val currentTime = Clock.System.now()